+++ Sources/Yams/Node.Sequence.swift
@@ -18,6 +18,8 @@ extension Node {
         public var style: Style
         /// The location for this node.
         public var mark: Mark?
+        /// The anchor for this node.
+        public weak var anchor: Anchor?
 
         /// The style to use when emitting a `Sequence`.
         public enum Style: UInt32 {
@@ -35,11 +37,12 @@ extension Node {
         /// - parameter tag:   This sequence's `Tag`.
         /// - parameter style: The style to use when emitting this `Sequence`.
         /// - parameter mark:  This sequence's `Mark`.
-        public init(_ nodes: [Node], _ tag: Tag = .implicit, _ style: Style = .any, _ mark: Mark? = nil) {
+        public init(_ nodes: [Node], _ tag: Tag = .implicit, _ style: Style = .any, _ mark: Mark? = nil, _ anchor: Anchor? = nil) {
             self.nodes = nodes
             self.tag = tag
             self.style = style
             self.mark = mark
+            self.anchor = anchor
         }
     }